# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "TurboStream API for real-time data streaming and feed management with AI-powered analysis",
        "title": "TurboStream API",
        "termsOfService": "https://turboline.ai/terms",
        "contact": {
            "name": "API Support",
            "url": "https://turboline.ai/docs/api",
            "email": "dev@turboline.ai"
        },
        "license": {
            "name": "MPL-2.0",
            "url": "https://github.com/turboline-ai/turbostream?tab=MPL-2.0-1-ov-file"
        },
        "version": "1.0"
    },
    "host": "localhost:7210",
    "basePath": "/",
    "paths": {
...
        "/api/llm/query": {
            "post": {
                "security": [{"BearerAuth": []}],
                "description": "Query feed data with an LLM. Returns complete response with token usage.",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["LLM"],
                "summary": "Query feed with LLM",
                "parameters": [
                    {
                        "description": "Query details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/LLMQueryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/LLMQueryResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/llm/query/stream": {
            "post": {
                "security": [{"BearerAuth": []}],
                "description": "Stream LLM responses using Server-Sent Events (SSE). Returns tokens as they are generated.",
                "consumes": ["application/json"],
                "produces": ["text/event-stream"],
                "tags": ["LLM"],
                "summary": "Stream LLM query",
                "parameters": [
                    {
                        "description": "Query details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/LLMQueryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK - Returns SSE stream of tokens"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "LLMQueryRequest": {
            "type": "object",
            "required": ["feedId", "question"],
            "properties": {
                "feedId": {
                    "type": "string",
                    "example": "507f1f77bcf86cd799439011"
                },
                "question": {
                    "type": "string",
                    "example": "What is the current Bitcoin price trend?"
                },
                "provider": {
                    "type": "string",
                    "enum": ["openai", "anthropic", "google", "mistral", "xai", "ollama", "azure"],
                    "example": "anthropic"
                },
                "systemPrompt": {
                    "type": "string",
                    "description": "Custom system prompt (optional)",
                    "example": "You are a cryptocurrency trading analyst."
                }
            }
        },
        "LLMQueryResponse": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string",
                    "example": "Based on the recent data, Bitcoin is showing an upward trend..."
                },
                "provider": {
                    "type": "string",
                    "example": "anthropic"
                },
                "feedId": {
                    "type": "string",
                    "example": "507f1f77bcf86cd799439011"
                },
                "tokensUsed": {
                    "type": "integer",
                    "example": 150
                },
                "durationMs": {
                    "type": "integer",
                    "example": 1250
                }
            }
        },
...
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean",
                    "example": false
                },
                "message": {
                    "type": "string",
                    "example": "Error message describing what went wrong"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type 'Bearer' followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}